# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Service API v2
  description: |
    물품 배송 서비스 백엔드 API 명세서

    ## Architecture
    - Hexagonal Architecture (Ports & Adapters)
    - Spring Boot 4.0.1 / Java 21
    - Event-Driven Architecture

    ## Service Type
    - Product Delivery Service (물품 배송 서비스)
    - Courier-based delivery (택배사 연동)
    - Domestic shipping only (국내 배송)
    - Return/Exchange supported (반품/교환 지원)

    ## Domains
    1. Customer - 고객 관리
    2. Product - 상품 관리
    3. Seller - 판매자 관리
    4. Cart - 장바구니
    5. Order - 주문 관리
    6. Payment - 결제 관리
    7. Shipment - 배송 관리
    8. Return - 반품/교환 관리
    9. Review - 리뷰 관리
    10. Promotion - 쿠폰/프로모션
  version: 2.0.0
  contact:
    name: Delivery Service Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v2
    description: Local Development Server
  - url: https://api.delivery.com/api/v2
    description: Production Server
...
paths:
...
  /sellers/{sellerId}:
    get:
      tags: [Sellers]
      summary: 판매자 정보 조회
      operationId: getSeller
      parameters:
        - $ref: '#/components/parameters/SellerId'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SellerResponse'
...
  /shipments/{shipmentId}:
    get:
      tags: [Shipments]
      summary: 배송 정보 조회
      operationId: getShipment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShipmentId'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentResponse'
  /shipments/{shipmentId}/tracking:
    get:
      tags: [Shipments]
      summary: 배송 추적 조회
      operationId: getShipmentTracking
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShipmentId'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipmentTrackingResponse'

  # ============================================
  # Return API
  # ============================================
...
  /webhooks/shipments/{carrier}:
    post:
      tags: [Webhooks]
      summary: 배송 웹훅
      operationId: handleShipmentWebhook
      parameters:
        - name: carrier
          in: path
          required: true
          schema:
            type: string
            enum: [cj, hanjin, lotte, logen, post]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: 처리 완료
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SellerId:
      name: sellerId
      in: path
      required: true
      schema:
        type: string
...
    ShipmentId:
      name: shipmentId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ShipmentStatus:
      type: string
      enum:
        - PENDING
        - READY_TO_SHIP
        - PICKED_UP
        - IN_TRANSIT
        - OUT_FOR_DELIVERY
        - DELIVERED
        - FAILED
        - RETURNED
    ShippingCarrier:
      type: string
      enum:
        - CJ_LOGISTICS
        - HANJIN
        - LOTTE
        - LOGEN
        - POST_OFFICE
        - EPOST
        - GS_POSTBOX
        - CU_POST
        - OTHER
...
    SellerResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        sellerType:
          type: string
        ratingAvg:
          type: number
        reviewCount:
          type: integer
        productCount:
          type: integer
        status:
          type: string
...
    ShipmentResponse:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        carrier:
          $ref: '#/components/schemas/ShippingCarrier'
        carrierName:
          type: string
        trackingNumber:
          type: string
        trackingUrl:
          type: string
        status:
          $ref: '#/components/schemas/ShipmentStatus'
        estimatedDeliveryDate:
          type: string
          format: date
        deliveredAt:
          type: string
          format: date-time
...
    ShipmentTrackingResponse:
      type: object
      properties:
        shipmentId:
          type: string
        carrier:
          $ref: '#/components/schemas/ShippingCarrier'
        trackingNumber:
          type: string
        status:
          $ref: '#/components/schemas/ShipmentStatus'
        events:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              location:
                type: string
              description:
                type: string
              occurredAt:
                type: string
                format: date-time
...